requirements.txt
altair
numpy
pandas
pydeck
streamlit
requests
os
sklearn.linear_model
matplotlib.pyplot